P4ADEV-2626 moved class
