+++ src/test/java/it/gov/pagopa/payhub/activities/config/interceptor/QueryParamsPlusEncoderInterceptorTest.java
@@ -1,5 +1,6 @@
 package it.gov.pagopa.payhub.activities.config.interceptor;
 
+import it.gov.pagopa.payhub.activities.config.rest.QueryParamsPlusEncoderInterceptor;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
